视频测试完成后跳转到当前测试结果页面。支持再测一次

+++ SPUIView/SPService/SVVideoTestResult.h
@@ -28,7 +28,7 @@
 
 
 // 首次缓冲时间
-@property long firstBufferTime;
+@property int firstBufferTime;
 
 // 开始视频播放
 @property long videoStartPlayTime;
